Fixed build

+++ Projects/Engine/Framework/RotateSystem2D.cpp
@@ -29,7 +29,7 @@ void RotateSystem2D::Run(Components* pComp, const float fDelta)
 	if(fDiff > s_fMinDiff)
 	{
 		//float dir = Mathf.Sign(((((target - current) % 360) + 540) % 360) - 180);    //  Get the shortest angle
-		float fDir = math::Sign((((static_cast<int>(rot.target - trans.fAngle) % 360) + 540) % 360) - 180);	// get the shortest direction to target
+		float fDir = (float)math::Sign((((static_cast<int>(rot.target - trans.fAngle) % 360) + 540) % 360) - 180);	// get the shortest direction to target
 		float fSpeed = rot.speed * fDelta;
 		float fAngle = math::RotateAngle(trans.fAngle, (fDir * fSpeed) * (fDiff < s_fLean ? (fDiff / s_fLean) : 1.0f));
 		fAngle = math::Clamp(fAngle, rot.min, rot.max);
